Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-invoke-in-sync-event-expected.txt


FAIL await a stable state and sync event handlers assert_equals: networkState in separate script expected 0 but got 3
